+++ src/test/java/autotests/tests/cart/CalculateTotalTests.java
@@ -23,8 +23,8 @@ public class CalculateTotalTests extends TestBase {
 
     @Test(dataProvider = "items", dataProviderClass = DataProviders.class)
     public void testCalculateTotal(ItemData firstItem, ItemData secondItem, ItemData thirdItem){
-        float costSum = app.costSumOfItems();
-        float total = app.totalOnTheCartPage();
+        double costSum = app.costSumOfItems();
+        double total = app.totalOnTheCartPage();
 
         assertEquals(total, costSum);
     }
